Split SFX mutations into per-domain sub-routers
Addresses the qlty file-complexity smell on sfxMutations.ts by splitting
its nine handlers into domain-focused routers — triggers, sound files and
categories — mounted together by a thin aggregator, mirroring the existing
commandMutations / commandAssignments split. Shared file-deletion logic
moves to sfxMutationsShared. Behaviour is unchanged; server wiring still
references the aggregator's default export.

Helper unit tests (detectAudioType, buildStoredName) move alongside their
new home in sfxFileMutations; route integration tests continue to exercise
the aggregator.
